Don't refer to related docs that no longer exist
The query would return doc ids related to the one the user is
accessing, even for docs that no longer exist.
The bug was being triggered on the next line of code, when we instantiate
each of those related docs.
Added a filter to the query to list only existing related docs.
As I'm not 100% aware of what the relateddocs is used for, I didn't
consider deleting the records pointing to the inexistent one. Still, if
that was the case it should probably be deleted at the time that doc is
deleted.